fix: fixed slug page
