+++ pages/men/[slug].tsx
@@ -12,6 +12,7 @@ import { Product, Sku } from '../../utils/helpers/types/product';
 import Image from 'next/image';
 import useBasket from '../../store/basket';
 import useHasHydrated from '../../components/UseHasHydrated';
+import FullProduct from '../../components/ProductsView/FullProduct';
 
 interface IParams extends ParsedUrlQuery {
   slug: string;
@@ -32,6 +33,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
 
 export const getStaticProps = async ({ params }: { params: IParams }) => {
   const product = await ContentfulApi.getProductBySlug(`/${params.slug}`);
+  const sku = await ContentfulApi.getProductSkus(product.sys.id);
   if (!product) {
     return {
       notFound: true,
@@ -40,124 +42,24 @@ export const getStaticProps = async ({ params }: { params: IParams }) => {
   return {
     props: {
       product,
+      sku,
     },
   };
 };
 
-const Page = ({ product }: { product: Product }) => {
-  const [selectedSize, setSelectedSize] = useState<Sku | undefined>(undefined);
-  const [selectedQuantity, setSelectedQuantity] = useState<number>(1);
-  const hasHydrated = useHasHydrated();
-
-  const { addToBasket } = useBasket();
-
+const Page = ({ product, sku }: { product: Product; sku: Sku[] }) => {
   if (!product) return <>Loading...</>;
 
-  const { name, description, price, sys, rating, imagesCollection, department, sku } =
-    product;
   return (
     <Layout
       seo={{
-        title: `${name} | QuackStore`,
+        title: `${product.name} | QuackStore`,
         description: `Mens | QuackStore`,
         canonicalUrl: `https://quackstore.com/men`,
-        image: imagesCollection.items[0].url,
+        image: product.imagesCollection.items[0].url,
       }}
     >
-      <section className="wrapper mt-32">
-        <article className="w-1/2 px-4">
-          <figure>
-            <Image
-              src={imagesCollection.items[0].url}
-              layout="responsive"
-              width={imagesCollection.items[0].width}
-              height={imagesCollection.items[0].height}
-              alt={imagesCollection.items[0].title}
-            />
-          </figure>
-        </article>
-        <article className="w-1/2 flex flex-col gap-3">
-          <div className="w-full flex justify-between">
-            <p>{department}</p>
-            <Rating rating={rating} />
-          </div>
-          <h1>{product.name}</h1>
-          <div className="flex gap-5">
-            <p>Available in:</p>
-            {sku?.map(item => {
-              return (
-                <button
-                  key={`${sys.id}${item.colour}`}
-                  type="button"
-                  className={clsx(
-                    'rounded-full h-5 w-5 border',
-                    `bg-${item.colour.toLocaleLowerCase()}`
-                  )}
-                  aria-label="item color"
-                />
-              );
-            })}
-          </div>
-          <h2>£{price}</h2>
-
-          <h3>Select size</h3>
-          <div className="w-full">
-            <RadioSelect
-              id={'size'}
-              name={'Size'}
-              value={selectedSize}
-              options={sku}
-              type={'size'}
-              onChange={option => setSelectedSize(option)}
-            />
-          </div>
-          <div className="flex justify-between mt-5">
-            <div className="flex flex-row">
-              <button
-                type="button"
-                className="bg-grey-200 w-12 h-12 border-y border-l"
-                onClick={() => {
-                  selectedQuantity > 1 && setSelectedQuantity(selectedQuantity - 1);
-                }}
-              >
-                -
-              </button>
-              <p className="flex items-center justify-center w-14 h-12 border-y">
-                {selectedQuantity}
-              </p>
-              <button
-                type="button"
-                className="bg-grey-200 w-12 h-12 border-y border-r"
-                onClick={() => {
-                  setSelectedQuantity(selectedQuantity + 1);
-                }}
-              >
-                +
-              </button>
-            </div>
-            <Button
-              type="primary"
-              onClick={() =>
-                hasHydrated &&
-                addToBasket({
-                  ...product,
-                  quantity: selectedQuantity,
-                })
-              }
-              disabled={!selectedSize}
-            >
-              Add to cart
-            </Button>
-          </div>
-          <div className="text-granite-gray mt-10">
-            <p className="text-primary font-semibold text-lg">Product Info</p>
-            {
-              // @ts-ignore
-              documentToReactComponents(description.json)
-            }
-          </div>
-        </article>
-      </section>
+      <FullProduct product={product} sku={sku} />
     </Layout>
   );
 };
